requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
tiktoken>=0.5.0

# Data processing
datasets>=2.14.0
transformers>=4.30.0
torch>=2.0.0

# Analysis and visualization  
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
pyyaml>=6.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
